builds /update_user route
